requirements.txt
# LLM Frameworks
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# LlamaIndex
llama-index>=0.11.0
llama-index-readers-file>=0.4.0
llama-index-llms-openai>=0.3.0
llama-index-llms-openai-like>=0.2.0
llama-index-embeddings-openai>=0.3.0
llama-index-embeddings-huggingface>=0.3.0
sentence-transformers>=3.0.0

# PDF reading (for resume)
pypdf>=4.0.0

# Web
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12

# Storage
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Speech-to-text
funasr>=1.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
rich>=13.0.0
